requirements.txt
numpy==1.18.1
pytorch-crf==0.7.2
tensorboard==2.1.0
torch==1.4.0
torchvision==0.5.0
transformers==2.3.0
sklearn==0.0